+++ Common/Image/Image.h
@@ -1,7 +1,7 @@
 #ifndef _CC_IMAGE_H
 #define _CC_IMAGE_H
 
-#include "../Colour/RGB.h"
+#include "RGB.h"
 
 #include <cstdint>
 #include <vector>
@@ -22,53 +22,53 @@ namespace cc
 	class Image final
 	{
 	public:
-		Image(size_t width, size_t height);
-		Image(size_t width, size_t height, const RGBA& fillColour);
+		Image(int32_t width, int32_t height);
+		Image(int32_t width, int32_t height, const RGBA& fillColour);
 		Image(const Image8& img8);
 
-		RGBA& operator() (size_t x, size_t y);
-		const RGBA& operator() (size_t x, size_t y) const;
+		RGBA& operator() (int32_t x, int32_t y);
+		const RGBA& operator() (int32_t x, int32_t y) const;
 
-		RGBA& operator[] (size_t pixelIndex);
-		const RGBA& operator[] (size_t pixelIndex) const;
+		RGBA& operator[] (int32_t pixelIndex);
+		const RGBA& operator[] (int32_t pixelIndex) const;
 
-		size_t Width() const;
-		size_t Height() const;
-		size_t PixelCount() const;
+		int32_t Width() const;
+		int32_t Height() const;
+		int32_t PixelCount() const;
 
 		static std::optional<Image> Load(std::string_view filepath);
 		static void  Save(std::string_view filepath, const Image& image, ImageType type);
 
 	private:
 		std::vector<RGBA> m_Pixels;
-		size_t m_Width;
-		size_t m_Height;
+		int32_t m_Width;
+		int32_t m_Height;
 	};
 
 	class Image8 final
 	{
 	public:
-		Image8(size_t width, size_t height);
-		Image8(size_t width, size_t height, const RGBA8& fillColour);
+		Image8(int32_t width, int32_t height);
+		Image8(int32_t width, int32_t height, const RGBA8& fillColour);
 		Image8(const Image& img);
 
-		RGBA8& operator() (size_t x, size_t y);
-		const RGBA8& operator() (size_t x, size_t y) const;
+		RGBA8& operator() (int32_t x, int32_t y);
+		const RGBA8& operator() (int32_t x, int32_t y) const;
 
-		RGBA8& operator[] (size_t pixelIndex);
-		const RGBA8& operator[] (size_t pixelIndex) const;
+		RGBA8& operator[] (int32_t pixelIndex);
+		const RGBA8& operator[] (int32_t pixelIndex) const;
 
-		size_t Width() const;
-		size_t Height() const;
-		size_t PixelCount() const;
+		int32_t Width() const;
+		int32_t Height() const;
+		int32_t PixelCount() const;
 
 		static std::optional<Image8> Load(std::string_view filepath);
 		static void                  Save(std::string_view filepath, const Image8& image, ImageType type);
 
 	private:
 		std::vector<RGBA8> m_Pixels;
-		size_t m_Width;
-		size_t m_Height;
+		int32_t m_Width;
+		int32_t m_Height;
 	};
 }
